+++ src/type-extensions.ts
@@ -2,11 +2,7 @@ import "ethers";
 import "hardhat/types/config";
 import "hardhat/types/runtime";
 
-import { Migrator } from "./migrator/Migrator";
-
 import { MigrateConfig } from "./types/migrations";
-
-import { MigrateStorage } from "./tools/storage/MigrateStorage";
 
 declare module "hardhat/types/config" {
   interface HardhatConfig {
@@ -15,13 +11,6 @@ declare module "hardhat/types/config" {
 
   interface HardhatUserConfig {
     migrate?: Partial<MigrateConfig>;
-  }
-}
-
-declare module "hardhat/types/runtime" {
-  export interface HardhatRuntimeEnvironment {
-    migrator: Migrator;
-    storage: MigrateStorage;
   }
 }